Extract some shared helpers to codegen/types
