+++ internal/codegen/shared.go
@@ -5,12 +5,10 @@ import (
 	"io"
 	"os"
 	"path"
-	"strconv"
 	"strings"
 	"unicode"
 
 	"github.com/dave/jennifer/jen"
-	"github.com/ethanmoffat/eolib-go/internal/xml"
 )
 
 func getPackageStatement(outputDir string) (string, error) {
@@ -66,13 +64,6 @@ func sanitizeComment(comment string) string {
 	return strings.Join(split, " ")
 }
 
-func sanitizeTypeName(typeName string) string {
-	if strings.HasSuffix(typeName, "Type") {
-		return typeName[:len(typeName)-4]
-	}
-	return typeName
-}
-
 func writeTypeCommentJen(f *jen.File, typeName string, comment string) {
 	if comment = sanitizeComment(comment); len(comment) > 0 {
 		f.Commentf("// %s :: %s", typeName, comment)
@@ -124,55 +115,6 @@ func writeToFile(outFileName string, outputText string) error {
 	return nil
 }
 
-type importInfo struct {
-	Package string
-	Path    string
-}
-
-func eoTypeToGoType(eoType string, currentPackage string, fullSpec xml.Protocol) (goType string, nextImport *importInfo) {
-	if strings.ContainsRune(eoType, rune(':')) {
-		eoType = strings.Split(eoType, ":")[0]
-	}
-
-	switch eoType {
-	case "byte":
-		fallthrough
-	case "char":
-		fallthrough
-	case "short":
-		fallthrough
-	case "three":
-		fallthrough
-	case "int":
-		return "int", nil
-	case "bool":
-		return "bool", nil
-	case "blob":
-		return "[]byte", nil
-	case "string":
-		fallthrough
-	case "encoded_string":
-		return "string", nil
-	default:
-		match := fullSpec.FindType(eoType)
-		goType = eoType
-
-		if structMatch, ok := match.(*xml.ProtocolStruct); ok && structMatch.Package != currentPackage {
-			nextImport = &importInfo{structMatch.Package, structMatch.PackagePath}
-		} else if enumMatch, ok := match.(*xml.ProtocolEnum); ok && enumMatch.Package != currentPackage {
-			nextImport = &importInfo{enumMatch.Package, enumMatch.PackagePath}
-		}
-
-		if nextImport != nil {
-			if val, ok := packageAliases[nextImport.Package]; ok {
-				nextImport.Path = val
-			}
-		}
-
-		return
-	}
-}
-
 func snakeCaseToCamelCase(input string) string {
 	if len(input) == 0 {
 		return input
@@ -209,137 +151,4 @@ func snakeCaseToPascalCase(input string) string {
 	camelCase := snakeCaseToCamelCase(input)
 	firstRune := []rune(camelCase)[0]
 	return string(unicode.ToUpper(firstRune)) + camelCase[1:]
-}
-
-func getInstructionTypeName(inst xml.ProtocolInstruction) (typeName string, typeSize string) {
-	if inst.Type == nil {
-		return
-	}
-
-	if strings.ContainsRune(*inst.Type, rune(':')) {
-		split := strings.Split(*inst.Type, ":")
-		typeName, typeSize = split[0], split[1]
-	} else {
-		typeName = *inst.Type
-	}
-
-	return
-}
-
-func calculateTypeSize(typeName string, fullSpec xml.Protocol) (res int, err error) {
-	var structInfo *xml.ProtocolStruct
-	var isStruct bool
-	if structInfo, isStruct = fullSpec.IsStruct(typeName); !isStruct {
-		return getPrimitiveTypeSize(typeName, fullSpec)
-	}
-
-	var flattenedInstList []xml.ProtocolInstruction
-	for _, instruction := range (*structInfo).Instructions {
-		if instruction.XMLName.Local == "chunked" {
-			flattenedInstList = append(flattenedInstList, instruction.Chunked...)
-		} else {
-			flattenedInstList = append(flattenedInstList, instruction)
-		}
-	}
-
-	for _, instruction := range flattenedInstList {
-		switch instruction.XMLName.Local {
-		case "field":
-			fieldTypeName, fieldTypeSize := getInstructionTypeName(instruction)
-			if fieldTypeSize != "" {
-				fieldTypeName = fieldTypeSize
-			}
-
-			if instruction.Length != nil {
-				if length, err := strconv.ParseInt(*instruction.Length, 10, 32); err == nil {
-					// length is a numeric constant
-					res += int(length)
-				} else {
-					return 0, fmt.Errorf("instruction length %s must be a fixed size for %s (%s)", *instruction.Length, *instruction.Name, instruction.XMLName.Local)
-				}
-			} else {
-				if nestedSize, err := getPrimitiveTypeSize(fieldTypeName, fullSpec); err != nil {
-					return 0, err
-				} else {
-					res += nestedSize
-				}
-			}
-		case "break":
-			res += 1
-		case "array":
-		case "dummy":
-		}
-	}
-
-	return
-}
-
-func getPrimitiveTypeSize(fieldTypeName string, fullSpec xml.Protocol) (int, error) {
-	switch fieldTypeName {
-	case "byte":
-		fallthrough
-	case "char":
-		return 1, nil
-	case "short":
-		return 2, nil
-	case "three":
-		return 3, nil
-	case "int":
-		return 4, nil
-	case "bool":
-		return 1, nil
-	case "blob":
-		fallthrough
-	case "string":
-		fallthrough
-	case "encoded_string":
-		return 0, fmt.Errorf("cannot get size of %s without fixed length", fieldTypeName)
-	default:
-		if _, isStruct := fullSpec.IsStruct(fieldTypeName); isStruct {
-			return calculateTypeSize(fieldTypeName, fullSpec)
-		} else if e, isEnum := fullSpec.IsEnum(fieldTypeName); isEnum {
-			enumTypeName := sanitizeTypeName(e.Type)
-			return getPrimitiveTypeSize(enumTypeName, fullSpec)
-		} else {
-			return 0, fmt.Errorf("cannot get fixed size of unrecognized type %s", fieldTypeName)
-		}
-	}
-}
-
-// structInfo is a type representing the metadata about a struct that should be rendered as generated code.
-// It represents the common properties of either a ProtocolPacket or a ProtocolStruct.
-type structInfo struct {
-	Name         string                    // Name is the name of the type. It is not converted from protocol naming convention (snake_case).
-	Comment      string                    // Comment is an optional type comment for the struct.
-	Instructions []xml.ProtocolInstruction // Instructions is a collection of instructions for the struct.
-	PackageName  string                    // PackageName is the containing package name for the struct.
-
-	Family                string // Family is the Packet Family of the struct, if the struct is a packet struct.
-	Action                string // Action is the Packet Action of the struct, if the struct is a packet struct.
-	SwitchStructQualifier string // SwitchStructQualifier is an additional qualifier prepended to structs used in switch cases in packets.
-}
-
-func getStructInfo(typeName string, fullSpec xml.Protocol) (si *structInfo, err error) {
-	si = &structInfo{SwitchStructQualifier: ""}
-	err = nil
-
-	if structInfo, ok := fullSpec.IsStruct(typeName); ok {
-		si.Name = structInfo.Name
-		si.Comment = structInfo.Comment
-		si.Instructions = structInfo.Instructions
-		si.PackageName = structInfo.Package
-	} else if packetInfo, ok := fullSpec.IsPacket(typeName); ok {
-		si.Name = packetInfo.GetTypeName()
-		si.Comment = packetInfo.Comment
-		si.Instructions = packetInfo.Instructions
-		si.PackageName = packetInfo.Package
-		si.SwitchStructQualifier = packetInfo.Family + packetInfo.Action
-		si.Family = packetInfo.Family
-		si.Action = packetInfo.Action
-	} else {
-		si = nil
-		err = fmt.Errorf("type %s is not a struct or packet in the spec", typeName)
-	}
-
-	return
 }
